requirements.txt
streamlit
faiss-cpu
sentence-transformers
google-generativeai
pandas
numpy
python-dotenv
